Renames queues to jobs
