+++ app/documentsAdapters/local/helpers/findOrCreateDocuments.ts
@@ -1,7 +1,7 @@
 import fse from 'fs-extra';
 import includes from 'lodash/includes.js';
 
-const COLLECTIONS = ['teams', 'users', 'prompts', 'promptVersions', 'projects', 'files', 'sessions', 'runs', 'collections', 'queues', 'featureFlags'];
+const COLLECTIONS = ['teams', 'users', 'prompts', 'promptVersions', 'projects', 'files', 'sessions', 'runs', 'collections', 'jobs', 'featureFlags'];
 
 const DEFAULTS = {
   teams: [],
@@ -13,7 +13,7 @@ const DEFAULTS = {
   collections: [],
   files: [],
   sessions: [],
-  queues: [],
+  jobs: [],
   featureFlags: []
 }
