+++ app/modules/documents/documents.ts
@@ -3,17 +3,17 @@ import 'app/documentsAdapters/documentDB/index.ts';
 import 'app/documentsAdapters/local/index.ts';
 
 import mongoose from 'mongoose';
-import projectSchema from './schemas/project.schema';
-import teamSchema from './schemas/team.schema';
-import userSchema from './schemas/user.schema';
-import promptSchema from './schemas/prompt.schema';
-import promptVersionSchema from './schemas/promptVersion.schema';
-import fileSchema from './schemas/file.schema';
-import sessionSchema from './schemas/session.schema';
-import runSchema from './schemas/run.schema';
 import collectionSchema from './schemas/collection.schema';
 import featureFlagSchema from './schemas/featureFlag.schema';
-import queueSchema from './schemas/queue.schema';
+import fileSchema from './schemas/file.schema';
+import jobSchema from './schemas/job.schema';
+import projectSchema from './schemas/project.schema';
+import promptSchema from './schemas/prompt.schema';
+import promptVersionSchema from './schemas/promptVersion.schema';
+import runSchema from './schemas/run.schema';
+import sessionSchema from './schemas/session.schema';
+import teamSchema from './schemas/team.schema';
+import userSchema from './schemas/user.schema';
 
 const registerModels = () => {
   if (!mongoose.models.Project) {
@@ -46,8 +46,8 @@ const registerModels = () => {
   if (!mongoose.models.FeatureFlag) {
     mongoose.model('FeatureFlag', featureFlagSchema);
   }
-  if (!mongoose.models.Queue) {
-    mongoose.model('Queue', queueSchema);
+  if (!mongoose.models.Job) {
+    mongoose.model('Job', jobSchema);
   }
 }
