+++ workers/helpers/localWorker.js
@@ -32,9 +32,9 @@ export default class LocalWorker {
   }
 
   setupQueue = () => {
-    const queuesPath = path.join(process.cwd(), `../data/queues.json`);
-    if (!fse.pathExistsSync(queuesPath)) {
-      fse.writeJsonSync(queuesPath, []);
+    const jobsPath = path.join(process.cwd(), `../data/jobs.json`);
+    if (!fse.pathExistsSync(jobsPath)) {
+      fse.writeJsonSync(jobsPath, []);
     }
   }
 
@@ -45,25 +45,25 @@ export default class LocalWorker {
         return;
       }
 
-      const queuesPath = path.join(process.cwd(), `../data/queues.json`);
-      const queues = await fse.readJson(queuesPath);
+      const jobsPath = path.join(process.cwd(), `../data/jobs.json`);
+      const jobs = await fse.readJson(jobsPath);
 
-      const jobs = filter(queues, (job) => {
+      const queueJobs = filter(jobs, (job) => {
         if (job.queue === this.name && job.attemptsMade < 3 && job.state === 'wait') {
           return job;
         }
       });
 
-      if (jobs.length > 0) {
+      if (queueJobs.length > 0) {
 
         this.isProcessing = true;
-        const currentJob = jobs[0];
+        const currentJob = queueJobs[0];
 
         try {
 
           currentJob.processedOn = new Date();
           currentJob.state = 'active';
-          await fse.writeJson(queuesPath, queues);
+          await fse.writeJson(jobsPath, jobs);
           this.emit('active', currentJob);
 
           await this.processJob(jobs[0]);
@@ -71,7 +71,7 @@ export default class LocalWorker {
           currentJob.finishedOn = new Date();
           currentJob.attemptsMade = jobs[0].attemptsMade + 1;
           currentJob.state = 'completed';
-          await fse.writeJson(queuesPath, queues);
+          await fse.writeJson(jobsPath, jobs);
           this.emit('completed', currentJob);
 
         } catch (error) {
@@ -88,7 +88,7 @@ export default class LocalWorker {
           } else {
             currentJob.state = 'wait';
           }
-          await fse.writeJson(queuesPath, queues);
+          await fse.writeJson(jobsPath, jobs);
           this.emit('failed', currentJob, error);
 
         }
@@ -109,17 +109,17 @@ export default class LocalWorker {
   cleanupJobs = async () => {
     try {
 
-      const queuesPath = path.join(process.cwd(), `../data/queues.json`);
-      const queues = await fse.readJson(queuesPath);
+      const jobsPath = path.join(process.cwd(), `../data/jobs.json`);
+      const jobs = await fse.readJson(jobsPath);
 
       // Removes queue items that are older than 30 days
-      remove(queues, (queue) => {
+      remove(jobs, (queue) => {
         if (dayjs(queue.timestamp).isBefore(dayjs().subtract(30, 'days'))) {
           return true;
         }
       });
 
-      await fse.writeJson(queuesPath, queues);
+      await fse.writeJson(jobsPath, jobs);
 
     } catch (error) {
       console.warn('Error during cleanupJobs', error);
